rename storybook-components
* `WithDocsCustom` now `withDocsCustom`
  * filename fixed too
* filename for `FinalFormWrapper` fixed

Closes #633
