+++ components/back-link/src/stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, text } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import BackLink from '.';
 import ReadMe from '../README.md';
@@ -10,7 +10,7 @@ import ReadMe from '../README.md';
 const stories = storiesOf('Navigation/BackLink', module);
 const examples = storiesOf('Navigation/BackLink/Examples', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 stories.addDecorator(withKnobs);
 
 stories.add('Component default', () => <BackLink goBack={action('go-back')}>{text('Children', 'Back')}</BackLink>);
