+++ components/breadcrumbs/src/stories.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import Breadcrumbs from '.';
 import ReadMe from '../README.md';
 
 const stories = storiesOf('Navigation/Breadcrumbs', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => (
   <Breadcrumbs>
