+++ components/button/src/stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import {
   ButtonWithKnobs,
@@ -18,7 +18,7 @@ const stories = storiesOf('Form/Buttons', module);
 const examples = storiesOf('Form/Buttons/Examples', module);
 
 stories.addDecorator(withKnobs);
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component Default', ButtonWithKnobs);
