+++ components/details/src/stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import { DetailsSimple, DetailsOpen, DetailsParagraph } from './fixtures';
 import ReadMe from '../README.md';
@@ -8,7 +8,7 @@ import ReadMe from '../README.md';
 const stories = storiesOf('Typography/Details', module);
 const examples = storiesOf('Typography/Details/Examples', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 stories.addDecorator(withKnobs);
 examples.addDecorator(withKnobs);
