+++ components/error-summary/src/stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import ErrorSummary from './fixtures';
 import ReadMe from '../README.md';
@@ -9,6 +9,6 @@ import ReadMe from '../README.md';
 const stories = storiesOf('Form/ErrorSummary', module);
 
 stories.addDecorator(withKnobs);
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => <ErrorSummary />);
