+++ components/grid-row/src/stories.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import styled from 'styled-components';
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 import GridCol from '@govuk-react/grid-col';
 
 import GridRow from '.';
@@ -9,7 +9,7 @@ import ReadMe from '../README.md';
 
 const stories = storiesOf('Layout/GridRow', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 const Content = styled('div')({
   textIndent: '-999em',
