+++ components/input-field/src/stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import InputField from '.';
 import ReadMe from '../README.md';
@@ -13,7 +13,7 @@ const meta = {
 const stories = storiesOf('Form/InputField', module);
 const examples = storiesOf('Form/InputField/Examples', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => <InputField name="group0">National Insurance number</InputField>);
