+++ components/list-item/src/stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 import Link from '@govuk-react/link';
 
 import ListItem from '.';
@@ -10,7 +10,7 @@ import ReadMe from '../README.md';
 const stories = storiesOf('Typography/ListItem', module);
 const examples = storiesOf('Typography/ListItem/Examples', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 stories.addDecorator(withKnobs);
 
 stories.add('Component default', () => <ListItem>{text('Children', 'List item example')}</ListItem>);
