+++ components/list-navigation/src/stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 import Link from '@govuk-react/link';
 
 import ListNavigation from '.';
@@ -8,7 +8,7 @@ import ReadMe from '../README.md';
 
 const stories = storiesOf('Navigation/ListNavigation', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => (
   <ListNavigation listStyleType="square">
