+++ components/main/src/stories.js
@@ -1,11 +1,11 @@
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import Main from '.';
 import ReadMe from '../README.md';
 
 const stories = storiesOf('Layout/Main', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', Main);
