+++ components/multi-choice/src/stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import MultiChoiceWithKnobs, { MultiChoiceWithKnobsHint, MultiChoiceWithKnobsError } from './fixtures';
 
@@ -8,7 +8,7 @@ import ReadMe from '../README.md';
 
 const stories = storiesOf('Form/MultiChoice', module);
 stories.addDecorator(withKnobs);
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 stories.add('Component default', MultiChoiceWithKnobs);
 
 const examples = storiesOf('Form/MultiChoice/Examples', module);
