+++ components/panel/src/stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import Panel from '.';
 import ReadMe from '../README.md';
@@ -8,7 +8,7 @@ import ReadMe from '../README.md';
 const stories = storiesOf('Misc/Panel', module);
 const examples = storiesOf('Misc/Panel/Examples', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => <Panel title="Application complete" />);
