+++ components/paragraph/src/stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import { SupportingParagraph, ParagraphWithKnobs } from './fixtures';
 import ReadMe from '../README.md';
@@ -9,7 +9,7 @@ const stories = storiesOf('Typography/Paragraph', module);
 const examples = storiesOf('Typography/Paragraph/Examples', module);
 
 stories.addDecorator(withKnobs);
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', ParagraphWithKnobs);
