+++ components/related-items/src/stories.js
@@ -4,14 +4,14 @@ import { H3 } from '@govuk-react/heading';
 import UnorderedList from '@govuk-react/unordered-list';
 import Link from '@govuk-react/link';
 import ListItem from '@govuk-react/list-item';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import RelatedItems from '.';
 import ReadMe from '../README.md';
 
 const stories = storiesOf('Navigation/RelatedItems', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => (
   <RelatedItems>
