+++ components/search-box/src/stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import Layout from '@govuk-react/layout';
 import GridRow from '@govuk-react/grid-row';
@@ -11,7 +11,7 @@ import ReadMe from '../README.md';
 
 const stories = storiesOf('Form/SearchBox', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => (
   <Layout>
