+++ components/skip-link/src/stories.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 import Paragraph from '@govuk-react/paragraph';
 
 import SkipLink from '.';
@@ -8,7 +8,7 @@ import SkipLink from '.';
 import ReadMe from '../README.md';
 
 const stories = storiesOf('Navigation/SkipLink', module);
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 class AutoFocussed extends Component {
   constructor(props) {
