+++ components/tag/src/stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import { TagWithKnobs, TagInactive } from './fixtures';
 import ReadMe from '../README.md';
@@ -8,7 +8,7 @@ import ReadMe from '../README.md';
 const stories = storiesOf('Misc/Tag', module);
 const examples = storiesOf('Misc/Tag/Examples', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 stories.addDecorator(withKnobs);
 examples.addDecorator(withKnobs);
