+++ components/text-area/src/stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import TextArea from '.';
 import ReadMe from '../README.md';
@@ -13,7 +13,7 @@ const meta = {
 const stories = storiesOf('Form/Textarea', module);
 const examples = storiesOf('Form/Textarea/Examples', module);
 
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => <TextArea name="group1">Description of what you saw</TextArea>);
