+++ components/warning-text/src/stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs/react';
-import { WithDocsCustom } from '@govuk-react/storybook-components';
+import { withDocsCustom } from '@govuk-react/storybook-components';
 
 import WarningText, { WarningWithLongText } from './fixtures';
 import ReadMe from '../README.md';
@@ -10,7 +10,7 @@ const stories = storiesOf('Typography/WarningText', module);
 const examples = storiesOf('Typography/WarningText/Examples', module);
 
 stories.addDecorator(withKnobs);
-stories.addDecorator(WithDocsCustom(ReadMe));
+stories.addDecorator(withDocsCustom(ReadMe));
 
 stories.add('Component default', () => <WarningText />);
