+++ packages/storybook-components/src/index.js
@@ -1,4 +1,4 @@
 /* eslint-disable import/prefer-default-export */
-export { default as FinalFormWrapper } from './FinalFormWrapper';
-export { default as WithDocsCustom } from './WithDocsCustom';
+export { default as FinalFormWrapper } from './final-form-wrapper';
+export { default as withDocsCustom } from './with-docs-custom';
 export { default as StoryPreview } from './story-preview';
